refactor: auth refreshToken dtos refactor
