--- src/auth/dtos/auth-login-output.dto.ts
@@ -1,3 +1,0 @@
-import { AuthToken } from './token.dto';
-
-export class LoginOutput extends AuthToken {}
