--- src/auth/dtos/auth-refresh-token-output.dto.ts
@@ -1,3 +1,0 @@
-import { AuthToken } from './token.dto';
-
-export class RefreshTokenOutput extends AuthToken {}
