+++ src/auth/services/auth.service.ts
@@ -6,10 +6,7 @@ import { UserService } from '../../user/services/user.service';
 import { User } from '../../user/entities/user.entity';
 import { RegisterInput } from '../dtos/auth-register-input.dto';
 import { RegisterOutput } from '../dtos/auth-register-output.dto';
-import { LoginOutput } from '../dtos/auth-login-output.dto';
-import { AuthToken } from '../dtos/token.dto';
-import { TokenUserIdentity } from '../dtos/token.dto';
-import { RefreshTokenOutput } from '../dtos/auth-refresh-token-output.dto';
+import { AuthToken, TokenUserIdentity } from '../dtos/auth-token-output.dto';
 
 @Injectable()
 export class AuthService {
@@ -29,23 +26,21 @@ export class AuthService {
     return user;
   }
 
-  login(user: User): LoginOutput {
-    return <LoginOutput>this.getAuthToken(user);
+  login(user: User): AuthToken {
+    return this.getAuthToken(user);
   }
 
   async register(input: RegisterInput): Promise<RegisterOutput> {
     return this.userService.createUser(input);
   }
 
-  async refreshToken(
-    tokenUser: TokenUserIdentity,
-  ): Promise<RefreshTokenOutput> {
+  async refreshToken(tokenUser: TokenUserIdentity): Promise<AuthToken> {
     const user = await this.userService.findById(tokenUser.id);
     if (!user) {
       throw new NotFoundException('Invalid user id');
     }
 
-    return <RefreshTokenOutput>this.getAuthToken(user);
+    return this.getAuthToken(user);
   }
 
   getAuthToken(user: { username: string; id: number }): AuthToken {
@@ -53,10 +48,10 @@ export class AuthService {
     const payload = { username: user.username, sub: user.id };
 
     return {
-      refresh_token: this.jwtService.sign(subject, {
+      refreshToken: this.jwtService.sign(subject, {
         expiresIn: this.configService.get('jwt.refreshTokenExpiresInSec'),
       }),
-      access_token: this.jwtService.sign(
+      accessToken: this.jwtService.sign(
         { ...payload, ...subject },
         { expiresIn: this.configService.get('jwt.accessTokenExpiresInSec') },
       ),
